#1495 Continued refactoring and added TODO´s with thoughts on the new structure
